+++ evals/tasks/t-003-rename-function/solution/app.py
@@ -1,0 +1,9 @@
+from greeting import new_greeting
+
+
+def greet_world():
+    return new_greeting("world")
+
+
+def greet_team():
+    return new_greeting("team")